+++ Classes/OldModel/Event.php
@@ -4606,20 +4606,10 @@ class Tx_Seminars_OldModel_Event extends \Tx_Seminars_OldModel_AbstractTimeSpan
                 'height' => $maxImageHeight,
             ],
         ];
-        $imageWithTag = $this->cObj->cObjGetSingle('IMAGE', $imageConfiguration);
+        $imageTag = $this->cObj->cObjGetSingle('IMAGE', $imageConfiguration);
 
-        $imageWidth = [];
-        preg_match('/width="([^"]*)"/', $imageWithTag, $imageWidth);
-        $imageHeight = [];
-        preg_match('/height="([^"]*)"/', $imageWithTag, $imageHeight);
-        $imageUrl = [];
-        preg_match('/src="([^"]*)"/', $imageWithTag, $imageUrl);
+        return $imageTag;
 
-        return ' style="' .
-            'background-image: url(\'' . $imageUrl[1] . '\'); ' .
-            'background-repeat: no-repeat; ' .
-            'padding-left: ' . $imageWidth[1] . 'px; ' .
-            'height: ' . $imageHeight[1] . 'px;"';
     }
 
     /**